Update DMXDetectorConstruction.cc
Removed the photo multiplier (pmt) so that the only object left is the LXe. To do this I had  to change "GXe_tube" to "LXe_tube" in the LXe_tube definition, deleted PMTdetector_tube definition, changed LXe_sol to LXe_tube in the LXe_log definition, changed the world_phys to lab_phys to in the LXe_phys definition, and uncommented (it was originally commented out) the line LXe_vat->SetForceSolid(true);. All of these changes were made in the liquid phase section. For the photomultiplier section, I deleted everything in it besides for the variable definitions (which I will come back to)