Update DMXDetectorConstruction.cc
I changed the liquid xenon into water, but it will still be called liquid xenon in the rest of the code as I only changed the material its made of. I changed the LXe_mat to H2O_mat in the LXe_log definition. And other changes had to be implemented in other files as well for this change to be successful