Update DMXEventAction.cc
Added additional branches to the third ntuple, and edited what branches were there originally. Edited the branches of xpos, ypos, and zpos so that they are accessing the ScintHitsCollection instead of the PMTHitsCollection. Added (and filled) the branches of TrackID, PDGCode, particleEnergy, hitEnergyDeposited, and hitTime. The TrackID and PDGCode required additional changes to other source and header files